Attempt 3, Revenge of the Sphere
Finally clicked that each test uses its own sets of variables based on the error message, so I can't use a variable defined in a different section of the code unless I redefine it in the section of code that the test runs.  Watch this now also not work and I'll have no clue where to go.